친구 추가 로직 수정

+++ controllers/user/friend/index.js
@@ -1,50 +1,76 @@
 import FriendService from "../../../services/user/friend/index.js"
 import User from "../../../models/user/User.js"
-
+import { sendNotification } from "../../../firebase/fcm.js"
 const friendService = new FriendService({
   User,
+  sendNotification,
 })
 
-export const addFriends = async (req, res) => {
+export const sendFriendRequest = async (req, res) => {
+  const userId = req.user.userId
+  const friendId = req.body.friendId
+
   try {
-    const userId = req.params.userId
-    const friendId = req.params.friendId
+    const result = await friendService.sendFriendRequest({ userId, friendId })
 
-    await friendService.addFriends({ userId, friendId })
+    return res.status(200).json({
+      success: true,
+      message: "친구 요청이 성공적으로 전송되었습니다.",
+      data: result,
+    })
+  } catch (error) {
+    return res.status(error.status || 500).json({
+      success: false,
+      message: error.message || "친구 요청 중 오류가 발생했습니다.",
+    })
+  }
+}
 
-    res.status(201).json({ message: "친구 추가 완료" })
-  } catch (err) {
-    console.log(err)
-    res.status(err.status || 500).json({ message: "서버 오류" })
+export const acceptFriendRequest = async (req, res) => {
+  const { userId, requesterId } = req.body
+
+  try {
+    const result = await friendService.acceptFriendRequest({
+      userId,
+      requesterId,
+    })
+    return res.status(200).json({
+      success: true,
+      message: "친구 요청을 수락했습니다.",
+      data: result,
+    })
+  } catch (error) {
+    return res.status(error.status || 500).json({
+      success: false,
+      message: error.message || "친구 요청 수락 중 오류가 발생했습니다.",
+    })
   }
 }
 
 export const removeFriends = async (req, res) => {
   try {
-    const userId = req.params.userId
-    const friendId = req.params.friendId
+    const { userId, friendId } = req.params
+    const isRemoved = await friendService.removeFriends({ userId, friendId })
 
-    await friendService.removeFriends({
-      userId,
-      friendId,
-    })
+    if (!isRemoved) {
+      return res.status(404).json({ message: "유저를 찾을 수 없음" })
+    }
 
-    res.status(200).json({ message: "친구 삭제 완료" })
+    return res.status(200).json({ message: "친구 삭제됨" })
   } catch (err) {
-    console.log(err)
-    res.status(err.status || 500).json({ message: "서버 오류" })
+    return res.status(500).json({ message: "서버 오류" })
   }
 }
 
 export const getFriendsCnt = async (req, res) => {
   try {
-    const userId = req.params.userId
+    const { userId } = req.params
+    const friends = await friendService.getFriendsCnt({ userId })
 
-    const cnt = await friendService.getFriendsCnt(userId)
-
-    res.status(200).json(cnt)
+    return res.status(200).json(friends)
   } catch (err) {
-    console.log(err)
-    res.status(err.status || 500).json({ message: "서버 오류" })
+    return res
+      .status(err.status || 500)
+      .json({ message: err.message || "서버 오류" })
   }
 }
